Protección con Middleware y Policies

+++ app/Http/Controllers/PermissionController.php
@@ -5,14 +5,19 @@ namespace App\Http\Controllers;
 use App\Models\Permission;
 use App\Http\Requests\Permission\StoreRequest;
 use App\Http\Requests\Permission\UpdateRequest;
-use Illuminate\Http\Request;
 use App\Models\Role;
 
 class PermissionController extends Controller
 {
 
+    public function __construct()
+    {
+        $this->middleware('permission:' . config('app.admin_permission'));
+    }
+
     public function index()
     {
+        $this->authorize('index',Permission::class);
         return view('theme.backoffice.pages.permission.index', [
             'permissions' => Permission::all()
         ]);
@@ -20,6 +25,7 @@ class PermissionController extends Controller
 
     public function create()
     {
+        $this->authorize('create',Permission::class);
         return view ('theme.backoffice.pages.permission.create',[
             'roles' => Role::all(),
         ]);
@@ -33,6 +39,7 @@ class PermissionController extends Controller
 
     public function show(Permission $permission)
     {
+        $this->authorize('view', $permission);
         return view('theme.backoffice.pages.permission.show', [
             'permission' => $permission
         ]);
@@ -40,6 +47,7 @@ class PermissionController extends Controller
 
     public function edit(Permission $permission)
     {
+        $this->authorize('update', $permission);
         return view('theme.backoffice.pages.permission.edit',[
             'permission' => $permission,
             'roles' => Role::all(),
@@ -54,6 +62,7 @@ class PermissionController extends Controller
 
     public function destroy(Permission $permission)
     {
+        $this->authorize('delete', $permission);
         $role = $permission->role;
         $permission->delete();
         toast('Permiso eliminado','success','top-right');
